refactor(screenshot): move dominantRatio out of the placeholder component
